Stable version of dht and fixing some bugs

+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/NettyKademliaDHTNode.java
@@ -1,25 +1,24 @@
 package io.Adrestus.p2p.kademlia;
 
 import io.Adrestus.p2p.kademlia.common.NettyConnectionInfo;
-import io.Adrestus.p2p.kademlia.server.KademliaNodeServer;
 import io.Adrestus.p2p.kademlia.node.DHTKademliaNodeAPI;
 import io.Adrestus.p2p.kademlia.node.DHTKademliaNodeAPIDecorator;
 import io.Adrestus.p2p.kademlia.node.Node;
+import io.Adrestus.p2p.kademlia.server.KademliaNodeServer;
 import lombok.Getter;
 import lombok.SneakyThrows;
 
 import java.io.Serializable;
 import java.math.BigInteger;
-import java.util.Date;
 import java.util.concurrent.Future;
 
 public class NettyKademliaDHTNode<K extends Serializable, V extends Serializable>
-        extends DHTKademliaNodeAPIDecorator<Long, NettyConnectionInfo, K, V> {
+        extends DHTKademliaNodeAPIDecorator<BigInteger, NettyConnectionInfo, K, V> {
 
     @Getter
     private final transient KademliaNodeServer<K, V> kademliaNodeServer;
 
-    public NettyKademliaDHTNode(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> kademliaNode, KademliaNodeServer<K, V> kademliaNodeServer) {
+    public NettyKademliaDHTNode(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> kademliaNode, KademliaNodeServer<K, V> kademliaNodeServer) {
         super(kademliaNode);
         this.kademliaNodeServer = kademliaNodeServer;
     }
@@ -33,7 +32,7 @@ public class NettyKademliaDHTNode<K extends Serializable, V extends Serializable
 
     @Override
     @SneakyThrows
-    public Future<Boolean> start(Node<Long, NettyConnectionInfo> bootstrapNode) {
+    public Future<Boolean> start(Node<BigInteger, NettyConnectionInfo> bootstrapNode) {
         kademliaNodeServer.run(this);
         return super.start(bootstrapNode);
     }
@@ -51,5 +50,4 @@ public class NettyKademliaDHTNode<K extends Serializable, V extends Serializable
         super.stopNow();
         kademliaNodeServer.stop();
     }
-
 }
