+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/common/NettyExternalNode.java
@@ -3,14 +3,16 @@ package io.Adrestus.p2p.kademlia.common;
 import io.Adrestus.p2p.kademlia.node.Node;
 import lombok.*;
 
+import java.math.BigInteger;
+
 
 @Data
 @ToString
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class NettyExternalNode implements Node<Long, NettyConnectionInfo> {
+public class NettyExternalNode implements Node<BigInteger, NettyConnectionInfo> {
     private NettyConnectionInfo connectionInfo;
-    private Long id;
+    private BigInteger id;
 
 }
