+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/factory/KademliaMessageHandlerFactory.java
@@ -1,16 +1,18 @@
 package io.Adrestus.p2p.kademlia.factory;
 
 import io.Adrestus.p2p.kademlia.common.NettyConnectionInfo;
+import io.Adrestus.p2p.kademlia.node.DHTKademliaNodeAPI;
 import io.Adrestus.p2p.kademlia.server.NettyKademliaMessageHandler;
 import io.Adrestus.p2p.kademlia.server.filter.NettyKademliaServerFilterChain;
-import io.Adrestus.p2p.kademlia.node.DHTKademliaNodeAPI;
 import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.handler.codec.http.FullHttpRequest;
 
 import java.io.Serializable;
+import java.math.BigInteger;
+
 
 public interface KademliaMessageHandlerFactory<K extends Serializable, V extends Serializable> {
-    SimpleChannelInboundHandler<FullHttpRequest> getKademliaMessageHandler(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI);
+    SimpleChannelInboundHandler<FullHttpRequest> getKademliaMessageHandler(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI);
 
     class DefaultKademliaMessageHandlerFactory<K extends Serializable, V extends Serializable> implements KademliaMessageHandlerFactory<K, V> {
         private final NettyKademliaServerFilterChain<K, V> filterChain;
@@ -20,7 +22,7 @@ public interface KademliaMessageHandlerFactory<K extends Serializable, V extends
         }
 
         @Override
-        public SimpleChannelInboundHandler<FullHttpRequest> getKademliaMessageHandler(DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> dhtKademliaNodeAPI) {
+        public SimpleChannelInboundHandler<FullHttpRequest> getKademliaMessageHandler(DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> dhtKademliaNodeAPI) {
             return new NettyKademliaMessageHandler<>(dhtKademliaNodeAPI, this.filterChain);
         }
