+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/serialization/DHTLookUpDataDeserializer.java
@@ -8,13 +8,14 @@ import io.Adrestus.p2p.kademlia.protocol.message.DHTLookupKademliaMessage;
 
 import java.io.Serializable;
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 
 
-public class DHTLookUpDataDeserializer<K extends Serializable> implements JsonDeserializer<DHTLookupKademliaMessage.DHTLookup<Long, NettyConnectionInfo, K>> {
+public class DHTLookUpDataDeserializer<K extends Serializable> implements JsonDeserializer<DHTLookupKademliaMessage.DHTLookup<BigInteger, NettyConnectionInfo, K>> {
 
     @Override
-    public DHTLookupKademliaMessage.DHTLookup<Long, NettyConnectionInfo, K> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-        DHTLookupKademliaMessage.DHTLookup<Long, NettyConnectionInfo, K> dhtLookup = new DHTLookupKademliaMessage.DHTLookup<>();
+    public DHTLookupKademliaMessage.DHTLookup<BigInteger, NettyConnectionInfo, K> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+        DHTLookupKademliaMessage.DHTLookup<BigInteger, NettyConnectionInfo, K> dhtLookup = new DHTLookupKademliaMessage.DHTLookup<>();
         JsonObject jsonObject = jsonElement.getAsJsonObject();
         JsonObject requesterJsonObject = jsonObject.getAsJsonObject("requester");
         dhtLookup.setRequester(jsonDeserializationContext.deserialize(requesterJsonObject, Node.class));
