+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/serialization/NodeDeserializer.java
@@ -6,14 +6,16 @@ import io.Adrestus.p2p.kademlia.common.NettyExternalNode;
 import io.Adrestus.p2p.kademlia.node.Node;
 
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 
-public class NodeDeserializer implements JsonDeserializer<Node<Long, NettyConnectionInfo>> {
+
+public class NodeDeserializer implements JsonDeserializer<Node<BigInteger, NettyConnectionInfo>> {
     @Override
-    public Node<Long, NettyConnectionInfo> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
+    public Node<BigInteger, NettyConnectionInfo> deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
         NettyExternalNode nettyExternalNode = new NettyExternalNode();
         JsonObject jsonObject = jsonElement.getAsJsonObject();
 
-        nettyExternalNode.setId(jsonDeserializationContext.deserialize(jsonObject.get("id"), Long.class));
+        nettyExternalNode.setId(jsonDeserializationContext.deserialize(jsonObject.get("id"), BigInteger.class));
         nettyExternalNode.setConnectionInfo(jsonDeserializationContext.deserialize(jsonObject.get("connectionInfo"), NettyConnectionInfo.class));
         return nettyExternalNode;
     }
