+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/serialization/NodeInstanceCreator.java
@@ -6,10 +6,12 @@ import io.Adrestus.p2p.kademlia.common.NettyExternalNode;
 import io.Adrestus.p2p.kademlia.node.Node;
 
 import java.lang.reflect.Type;
+import java.math.BigInteger;
 
-public class NodeInstanceCreator implements InstanceCreator<Node<Long, NettyConnectionInfo>> {
+
+public class NodeInstanceCreator implements InstanceCreator<Node<BigInteger, NettyConnectionInfo>> {
     @Override
-    public Node<Long, NettyConnectionInfo> createInstance(Type type) {
+    public Node<BigInteger, NettyConnectionInfo> createInstance(Type type) {
         return new NettyExternalNode();
     }
 }
