+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/server/filter/Context.java
@@ -5,17 +5,18 @@ import io.Adrestus.p2p.kademlia.node.DHTKademliaNodeAPI;
 import io.netty.channel.ChannelHandlerContext;
 
 import java.io.Serializable;
+import java.math.BigInteger;
+
 
 public interface Context<K extends Serializable, V extends Serializable> {
     ChannelHandlerContext getChannelHandlerContext();
-
-    DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> getDhtKademliaNodeApi();
+    DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> getDhtKademliaNodeApi();
 
     class ContextImpl<K extends Serializable, V extends Serializable> implements Context<K, V> {
         private final ChannelHandlerContext channelHandlerContext;
-        private final DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> getDhtKademliaNode;
+        private final DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> getDhtKademliaNode;
 
-        public ContextImpl(ChannelHandlerContext channelHandlerContext, DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> getDhtKademliaNode) {
+        public ContextImpl(ChannelHandlerContext channelHandlerContext, DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> getDhtKademliaNode) {
             this.channelHandlerContext = channelHandlerContext;
             this.getDhtKademliaNode = getDhtKademliaNode;
         }
@@ -26,7 +27,7 @@ public interface Context<K extends Serializable, V extends Serializable> {
         }
 
         @Override
-        public DHTKademliaNodeAPI<Long, NettyConnectionInfo, K, V> getDhtKademliaNodeApi() {
+        public DHTKademliaNodeAPI<BigInteger, NettyConnectionInfo, K, V> getDhtKademliaNodeApi() {
             return getDhtKademliaNode;
         }
     }
