+++ adrestus-network/src/main/java/io/Adrestus/p2p/kademlia/server/filter/KademliaMainHandlerFilter.java
@@ -1,7 +1,6 @@
 package io.Adrestus.p2p.kademlia.server.filter;
 
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
+
 import io.Adrestus.p2p.kademlia.common.NettyConnectionInfo;
 import io.Adrestus.p2p.kademlia.protocol.message.KademliaMessage;
 import io.Adrestus.p2p.kademlia.serialization.MessageSerializer;
@@ -11,10 +10,9 @@ import io.netty.handler.codec.http.FullHttpRequest;
 import io.netty.handler.codec.http.FullHttpResponse;
 import io.netty.util.CharsetUtil;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 
 import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_LENGTH;
@@ -34,10 +32,10 @@ public class KademliaMainHandlerFilter<K extends Serializable, V extends Seriali
 
     @Override
     public void filter(Context<K, V> context, FullHttpRequest request, FullHttpResponse response) {
-        KademliaMessage<Long, NettyConnectionInfo, ? extends Serializable> responseMessage = null;
+        KademliaMessage<BigInteger, NettyConnectionInfo, ? extends Serializable> responseMessage = null;
 
         try {
-            KademliaMessage<Long, NettyConnectionInfo, Serializable> kademliaMessage = this.toKademliaMessage(
+            KademliaMessage<BigInteger, NettyConnectionInfo, Serializable> kademliaMessage = this.toKademliaMessage(
                     this.parseJsonRequest(request)
             );
             responseMessage = context.getDhtKademliaNodeApi().onMessage(kademliaMessage);
@@ -63,7 +61,7 @@ public class KademliaMainHandlerFilter<K extends Serializable, V extends Seriali
         return jsonBuf.toString(CharsetUtil.UTF_8);
     }
 
-    protected KademliaMessage<Long, NettyConnectionInfo, Serializable> toKademliaMessage(String message) {
+    protected KademliaMessage<BigInteger, NettyConnectionInfo, Serializable> toKademliaMessage(String message) {
         return this.messageSerializer.deserialize(
                 message
         );
